# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "The API serves as the interface for APIs doc",
    "version": "1.0.0",
    "title": "FTRV Intranet Portal",
    "termsOfService": "",
    "contact": {
      "name": "Usman Qamar",
      "email": "usmanqamar189@hotmail.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Localhost FTRV Intranet"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/users": {
      "post": {
        "tags": ["User"],
        "summary": "An endpoint to create a user",
        "operationId": "createUsers",
        "requestBody": {
          "$ref": "#/components/requestBodies/User"
        },
        "responses": {
          "200": {
            "description": "Successful creation of user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. An Expert authenticated session needs to be created"
          }
        }
      },
      "get": {
        "summary": "Retrieve list of all user for given criteria",
        "tags": ["User"],
        "operationId": "getUserList",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pageNumber",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "searchString",
            "in": "query",
            "description": "Search string",
            "type": "string"
          },
          {
            "name": "sortColumn",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "departmentId",
            "in": "query",
            "type": "number"
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          },
          {
            "name": "extension",
            "in": "query",
            "type": "string"
          },
          {
            "name": "locationId",
            "in": "query",
            "type": "number"
          },
          {
            "name": "detail",
            "in": "query",
            "description": "True will return all user details false will return only name and id of users",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Retrieve user by id",
        "tags": ["User"],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. An Expert authenticated session needs to be created"
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "An endpoint to update a user",
        "operationId": "updateUsers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/User"
        },
        "responses": {
          "200": {
            "description": "Successful update of user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. An Expert authenticated session needs to be created"
          }
        }
      }
    },
    "/users/userBulkUpdate": {
      "put": {
        "tags": ["User"],
        "summary": "An endpoint to update bulk users",
        "operationId": "userBulkUpdate",
        "requestBody": {
          "$ref": "#/components/requestBodies/BulkUser"
        },
        "responses": {
          "200": {
            "description": "Successful update of user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUser"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. An Expert authenticated session needs to be created"
          }
        }
      }
    },
    "/users/deleteUsers": {
      "delete": {
        "tags": ["User"],
        "summary": "An endpoint to delete user(s)",
        "operationId": "deleteUsers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                },
                "required": ["ids"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful deletion of user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. An Expert authenticated session needs to be created"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["User"],
        "summary": "An endpoint to login",
        "operationId": "createUserSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful authentication of user credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Input"
          },
          "401": {
            "description": "Unauthorized. An Expert authenticated session needs to be created"
          }
        }
      }
    },
    "/users/upload": {
      "post": {
        "tags": ["User"],
        "summary": "An endpoint to upload employee sheet",
        "operationId": "uploadUserList",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "base64"
                  }
                },
                "required": ["file"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Count of successful and failed user creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/fileUpload"
                }
              }
            }
          },
          "400": {
            "description": "No file found!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Only excel file uploads are allowed!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/birthday": {
      "get": {
        "summary": "An endpoint to get user birthdays",
        "tags": ["User"],
        "operationId": "getBirthday",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Date for which we want to get birthdays",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/workAnniversary": {
      "get": {
        "summary": "An endpoint to get work anniversaries",
        "tags": ["User"],
        "operationId": "getAnniversary",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Date for which we want to get anniversaries",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkAnniversary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/groups/{id}/users": {
      "get": {
        "summary": "Retrieve list of all group users",
        "tags": ["Group"],
        "operationId": "getUserList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Group",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of Users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "requestBodies": {
      "User": {
        "description": "Information about the Users of FTRV",
        "content": {
          "multipart/form-data": {
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": ["user", "admin"]
                },
                "contactNo": {
                  "type": "string"
                },
                "extension": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "locationId": {
                  "type": "number"
                },
                "departmentId": {
                  "type": "number"
                },
                "file": {
                  "type": "string",
                  "format": "binary"
                },
                "joiningDate": {
                  "type": "string",
                  "format": "date"
                },
                "dob": {
                  "type": "string",
                  "format": "date"
                },
                "groupIds[]": {
                  "type": "array"
                }
              }
            }
          }
        },
        "required": true
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "",
        "scheme": "bearer",
        "name": "authorization",
        "bearerFormat": "JWT"
      }
    }
  }
}